Home (#24)
* [both] - feat: Add home doctor page

* [both] - feat: Add home patient page

* [frontend] - feat: Add home administrator page

* Adjust margin for search bar row in MedicalTestsList

Added an inline style to the row containing the search bar to set a negative top margin. This likely improves the visual alignment or spacing of the search bar within the component.

* Reorder fields in MedicalTestsList details view

Swapped the display order of 'Result' and 'Medical specialty' fields in the medical test details section for improved logical grouping.

* [backend] - test: Fix test_list_appointments_by_patient_with_patient